requirement.txt
asgiref==3.4.1
black==21.10b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.45.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
